Refactor typescript code
