+++ .github/workflows/js-build-test.yml
@@ -1,0 +1,35 @@
+name: Test JS build
+on:
+  pull_request:
+    branches:
+      - main
+  push:
+    branches:
+      - main
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    concurrency:
+      group: js-${{ github.ref }}
+      cancel-in-progress: true
+    steps:
+      - name: checkout
+        uses: actions/checkout@v2
+      - name: Use Node.js 14
+        uses: actions/setup-node@v2
+        with:
+          node-version: 14
+          cache: 'yarn'
+          cache-dependency-path: 'yarn.lock'
+      - name: Install node dependencies
+        run: yarn
+      - name: Build
+        run: yarn bob
+      - name: Lint
+        run: yarn lint
+      - name: Tests
+        run: yarn jest
+      - name: Build Example App
+        working-directory: Example/
+        run: yarn && yarn tsc
+
