+++ Example/src/examples/PanResponder.tsx
@@ -1,11 +1,11 @@
 import React, {PureComponent} from 'react';
 import {
-  PanResponder,
-  View,
   Animated,
+  PanResponder,
   TouchableWithoutFeedback,
+  View,
 } from 'react-native';
-import {Svg, G, Text, Path, Polyline, Line} from 'react-native-svg';
+import {G, Line, Path, Polyline, Svg, Text} from 'react-native-svg';
 
 const AnimatedSvg = Animated.createAnimatedComponent(Svg);
 
@@ -14,7 +14,7 @@ const zeroDelta = {x: 0, y: 0};
 class PanExample extends PureComponent {
   static title = 'Bind PanResponder on the SVG Shape';
   panXY: any;
-  constructor(props, context) {
+  constructor(props: {}, context: {}) {
     super(props, context);
     const xy = new Animated.ValueXY();
     const {x: dx, y: dy} = xy;
