+++ src/elements/Circle.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import { extract } from '../lib/extract/extractProps';
-import { NumberProp } from '../lib/extract/types';
+import { CommonPathProps, NumberProp } from '../lib/extract/types';
 import Shape from './Shape';
 import { RNSVGCircle } from './NativeComponents';
 
-export default class Circle extends Shape<{
+export interface CircleProps extends CommonPathProps {
   cx?: NumberProp;
   cy?: NumberProp;
+  opacity?: NumberProp;
   r?: NumberProp;
-}> {
+}
+
+export default class Circle extends Shape<CircleProps> {
   static displayName = 'Circle';
 
   static defaultProps = {
