+++ src/elements/ClipPath.tsx
@@ -1,9 +1,14 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { extract } from '../lib/extract/extractProps';
 import Shape from './Shape';
 import { RNSVGClipPath } from './NativeComponents';
 
-export default class ClipPath extends Shape<{}> {
+export interface ClipPathProps {
+  children?: ReactNode | undefined;
+  id?: string;
+}
+
+export default class ClipPath extends Shape<ClipPathProps> {
   static displayName = 'ClipPath';
 
   render() {
