+++ src/elements/Defs.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { RNSVGDefs } from './NativeComponents';
 
-export default class Defs extends Component {
+export default class Defs extends Component<React.PropsWithChildren<{}>> {
   static displayName = 'Defs';
 
   render() {
