+++ src/elements/Ellipse.tsx
@@ -1,15 +1,18 @@
 import React from 'react';
 import { extract } from '../lib/extract/extractProps';
-import { NumberProp } from '../lib/extract/types';
+import { CommonPathProps, NumberProp } from '../lib/extract/types';
 import Shape from './Shape';
 import { RNSVGEllipse } from './NativeComponents';
 
-export default class Ellipse extends Shape<{
+export interface EllipseProps extends CommonPathProps {
   cx?: NumberProp;
   cy?: NumberProp;
+  opacity?: NumberProp;
   rx?: NumberProp;
   ry?: NumberProp;
-}> {
+}
+
+export default class Ellipse extends Shape<EllipseProps> {
   static displayName = 'Ellipse';
 
   static defaultProps = {
