+++ src/elements/G.tsx
@@ -1,12 +1,22 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import extractProps, { propsAndStyles } from '../lib/extract/extractProps';
 import { extractFont } from '../lib/extract/extractText';
 import extractTransform from '../lib/extract/extractTransform';
-import { TransformProps } from '../lib/extract/types';
+import {
+  CommonPathProps,
+  FontProps,
+  NumberProp,
+  TransformProps,
+} from '../lib/extract/types';
 import Shape from './Shape';
 import { RNSVGGroup } from './NativeComponents';
 
-export default class G<P> extends Shape<P> {
+export interface GProps extends CommonPathProps, FontProps {
+  children?: ReactNode | undefined;
+  opacity?: NumberProp;
+}
+
+export default class G<P> extends Shape<GProps & P> {
   static displayName = 'G';
 
   setNativeProps = (
