+++ src/elements/Path.tsx
@@ -2,10 +2,14 @@ import React from 'react';
 import { extract } from '../lib/extract/extractProps';
 import Shape from './Shape';
 import { RNSVGPath } from './NativeComponents';
+import { CommonPathProps, NumberProp } from '../lib/extract/types';
 
-export default class Path extends Shape<{
+export interface PathProps extends CommonPathProps {
   d?: string;
-}> {
+  opacity?: NumberProp;
+}
+
+export default class Path extends Shape<PathProps> {
   static displayName = 'Path';
 
   render() {
