+++ src/elements/RadialGradient.tsx
@@ -1,23 +1,28 @@
 import React, { ReactElement } from 'react';
 import extractGradient from '../lib/extract/extractGradient';
-import { NumberProp, TransformProps } from '../lib/extract/types';
+import {
+  ColumnMajorTransformMatrix,
+  NumberProp,
+  Units,
+} from '../lib/extract/types';
 import Shape from './Shape';
 import { RNSVGRadialGradient } from './NativeComponents';
 
-export default class RadialGradient extends Shape<{
+export interface RadialGradientProps {
+  children?: ReactElement[];
   fx?: NumberProp;
   fy?: NumberProp;
   rx?: NumberProp;
   ry?: NumberProp;
-  r?: NumberProp;
   cx?: NumberProp;
   cy?: NumberProp;
+  r?: NumberProp;
+  gradientUnits?: Units;
+  gradientTransform?: ColumnMajorTransformMatrix | string;
   id?: string;
-  children?: ReactElement[];
-  transform?: number[] | string | TransformProps;
-  gradientTransform?: number[] | string | TransformProps;
-  gradientUnits?: 'objectBoundingBox' | 'userSpaceOnUse';
-}> {
+}
+
+export default class RadialGradient extends Shape<RadialGradientProps> {
   static displayName = 'RadialGradient';
 
   static defaultProps = {
