+++ src/elements/Rect.tsx
@@ -1,17 +1,20 @@
 import React from 'react';
 import { withoutXY } from '../lib/extract/extractProps';
-import { NumberProp } from '../lib/extract/types';
+import { CommonPathProps, NumberProp } from '../lib/extract/types';
 import Shape from './Shape';
 import { RNSVGRect } from './NativeComponents';
 
-export default class Rect extends Shape<{
+export interface RectProps extends CommonPathProps {
   x?: NumberProp;
   y?: NumberProp;
   width?: NumberProp;
   height?: NumberProp;
   rx?: NumberProp;
   ry?: NumberProp;
-}> {
+  opacity?: NumberProp;
+}
+
+export default class Rect extends Shape<RectProps> {
   static displayName = 'Rect';
 
   static defaultProps = {
