+++ src/elements/Use.tsx
@@ -1,18 +1,22 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { withoutXY } from '../lib/extract/extractProps';
-import { NumberProp } from '../lib/extract/types';
+import { CommonPathProps, NumberProp } from '../lib/extract/types';
 import { idPattern } from '../lib/util';
 import Shape from './Shape';
 import { RNSVGUse } from './NativeComponents';
 
-export default class Use extends Shape<{
-  x?: NumberProp;
-  y?: NumberProp;
-  width?: NumberProp;
-  height?: NumberProp;
+export interface UseProps extends CommonPathProps {
+  children?: ReactNode | undefined;
   xlinkHref?: string;
   href?: string;
-}> {
+  width?: NumberProp;
+  height?: NumberProp;
+  x?: NumberProp;
+  y?: NumberProp;
+  opacity?: NumberProp;
+}
+
+export default class Use extends Shape<UseProps> {
   static displayName = 'Use';
 
   static defaultProps = {
