+++ src/lib/extract/extractLengthList.ts
@@ -3,9 +3,16 @@ import { NumberProp } from './types';
 const spaceReg = /\s+/;
 const commaReg = /,/g;
 
+declare global {
+  interface ArrayConstructor {
+    // https://github.com/microsoft/TypeScript/issues/17002
+    isArray(arg: unknown): arg is unknown[] | readonly unknown[];
+  }
+}
+
 export default function extractLengthList(
-  lengthList?: NumberProp[] | NumberProp,
-): NumberProp[] {
+  lengthList?: readonly NumberProp[] | NumberProp,
+): readonly NumberProp[] {
   if (Array.isArray(lengthList)) {
     return lengthList;
   } else if (typeof lengthList === 'number') {
