+++ src/lib/extract/extractPolyPoints.ts
@@ -1,6 +1,15 @@
 import { NumberProp } from './types';
 
-export default function extractPolyPoints(points: string | NumberProp[]) {
+declare global {
+  interface ArrayConstructor {
+    // https://github.com/microsoft/TypeScript/issues/17002
+    isArray(arg: unknown): arg is unknown[] | readonly unknown[];
+  }
+}
+
+export default function extractPolyPoints(
+  points: string | readonly NumberProp[],
+) {
   const polyPoints = Array.isArray(points) ? points.join(',') : points;
   return polyPoints
     .replace(/[^eE]-/, ' -')
